+++ src/nexgen/beamlines/SSX_Tristan_nxs.py
@@ -112,11 +112,11 @@ def ssx_tristan_writer(
     logger.info(f"DLS Beamline: {beamline.upper()}.")
     if "I19" in beamline.upper():
         source = Source("I19-2")
-        from .beamline_parameters import I19_2Tristan as axes_params
+        from .beamline_utils import I19_2Tristan as axes_params
 
     elif "I24" in beamline.upper():
         source = Source("I19-2")
-        from .beamline_parameters import I24Eiger as axes_params
+        from .beamline_utils import I24Eiger as axes_params
 
         axes_params.fast_axis = Point3D(-1, 0, 0)
         axes_params.slow_axis = Point3D(0, 1, 0)
